Enlarge max job array task ID to 32-bits
The system really can not handle larger job arrays without adding
a job array data structure, but this puts some of the infrastructure
in place now.

+++ src/common/env.c
@@ -1115,7 +1115,7 @@ env_array_for_batch_job(char ***dest, const batch_job_launch_msg_t *batch,
 		env_array_overwrite_fmt(dest, "SLURM_BG_NUM_NODES",
 					"%u", num_nodes);
 	}
-	if (batch->array_task_id != (uint16_t) NO_VAL) {
+	if (batch->array_task_id != NO_VAL) {
 		env_array_overwrite_fmt(dest, "SLURM_ARRAY_JOB_ID", "%u",
 					batch->array_job_id);
 		env_array_overwrite_fmt(dest, "SLURM_ARRAY_TASK_ID", "%u",
